@@ -1,8 +1,11 @@
 package com.baeldung.jetty.httpclient;
 
 import org.eclipse.jetty.client.HttpClient;
-import org.junit.Assert;
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
 import org.springframework.http.client.reactive.ClientHttpConnector;
 import org.springframework.http.client.reactive.JettyClientHttpConnector;
@@ -12,25 +15,40 @@ import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Mono;
 
 public class SpringWebFluxUnitTest extends AbstractUnitTest {
-
+    
+    protected static int port = 9083;
+    
+    @BeforeAll
+    public static void init() {
+        startServer(new RequestHandler(), port);
+        startClient();
+    }
+    
     @Test
     public void givenReactiveClient_whenRequested_shouldReturnResponse() throws Exception {
-        
-        HttpClient httpClient = new HttpClient();
-        httpClient.start();
 
         ClientHttpConnector clientConnector = new JettyClientHttpConnector(httpClient);
         WebClient client = WebClient.builder()
             .clientConnector(clientConnector)
             .build();
         String responseContent = client.post()
-            .uri(uri())
+            .uri(uri(port))
             .contentType(MediaType.TEXT_PLAIN)
             .body(BodyInserters.fromPublisher(Mono.just(CONTENT), String.class))
             .retrieve()
             .bodyToMono(String.class)
             .block();
-        Assert.assertNotNull(responseContent);
-        Assert.assertEquals(CONTENT, responseContent);
+        assertNotNull(responseContent);
+        assertEquals(CONTENT, responseContent);
+    }
+    
+    @AfterAll
+    public static void dispose() throws Exception {
+        if (httpClient != null) {
+            httpClient.stop();
+        }
+        if (server != null) {
+            server.stop();
+        }
     }
 }
