@@ -4,7 +4,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class Spring5ReactiveApplication{
+public class Spring5ReactiveApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(Spring5ReactiveApplication.class, args);
