@@ -1,12 +1,9 @@
 package com.baeldung.properties;
 
-import com.baeldung.buildproperties.Application;
+import com.baeldung.configurationproperties.ConfigProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.context.annotation.ComponentScan;
-
-import com.baeldung.configurationproperties.ConfigProperties;
 
 @SpringBootApplication
 @ComponentScan(basePackageClasses = {ConfigProperties.class, AdditionalProperties.class})
