@@ -4,18 +4,16 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.servlet.http.HttpServletResponse;
-
 @RestController
 public class BaeldungController {
 
     @GetMapping("/hello")
-    public String sayHello(HttpServletResponse response) {
+    public String sayHello() {
         return "hello";
     }
 
     @PostMapping("/baeldung")
-    public String sayHelloPost(HttpServletResponse response) {
+    public String sayHelloPost() {
         return "hello";
     }
 }
