@@ -1,11 +1,11 @@
 package com.baeldung;
 
+import io.cucumber.junit.Cucumber;
+import io.cucumber.junit.CucumberOptions;
 import org.junit.runner.RunWith;
 
-import cucumber.api.CucumberOptions;
-import cucumber.api.junit.Cucumber;
 
 @RunWith(Cucumber.class)
 @CucumberOptions(features = "src/test/resources")
-public class CucumberIntegrationTest extends SpringIntegrationTest{
+public class CucumberIntegrationTest extends SpringIntegrationTest {
 }
