@@ -14,16 +14,18 @@ import org.springframework.http.client.ClientHttpResponse;
 
 public class LoggingInterceptor implements ClientHttpRequestInterceptor {
 
-    final static Logger log = LoggerFactory.getLogger(LoggingInterceptor.class);
+    final static Logger LOGGER = LoggerFactory.getLogger(LoggingInterceptor.class);
 
     @Override
     public ClientHttpResponse intercept(HttpRequest req, byte[] reqBody, ClientHttpRequestExecution ex) throws IOException {
-        log.debug("Request body: {}", new String(reqBody, StandardCharsets.UTF_8));
+        LOGGER.debug("Request body: {}", new String(reqBody, StandardCharsets.UTF_8));
         ClientHttpResponse response = ex.execute(req, reqBody);
-        InputStreamReader isr = new InputStreamReader(response.getBody(), StandardCharsets.UTF_8);
-        String body = new BufferedReader(isr).lines()
-            .collect(Collectors.joining("\n"));
-        log.debug("Response body: {}", body);
+        if (LOGGER.isDebugEnabled()) {
+            InputStreamReader isr = new InputStreamReader(response.getBody(), StandardCharsets.UTF_8);
+            String body = new BufferedReader(isr).lines()
+                .collect(Collectors.joining("\n"));
+            LOGGER.debug("Response body: {}", body);
+        }
         return response;
     }
 }
